backend-core/requirements/base.txt
responses==0.12.1
requests==2.25.0
authorizenet
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
celery[redis]==5.0.5
certifi==2020.11.8
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.2.1
daphne~=3.0.1
Django==3.1.6
django-websocket-redis==0.6.0
django_redis==4.12.1
git+git://github.com/ifdotpy-sielc/django-rest-framework-simplejwt@key-id#egg=djangorestframework_simplejwt
freezegun==1.0.0
gevent==20.12.1
google-api-python-client==1.12.8
djangorestframework==3.12.2
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
model_bakery==1.1.1
py4j==0.10.9.1
PyJWT==1.7.1
pytz==2020.4
psycopg2-binary==2.8.6
redis==3.5.3
sentry-sdk==0.20.1
six==1.15.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
django-jsonfield==1.4.1  # Need here for old migrations
tblib==1.7.0

-r requirements-peak.txt
